server/requirements.txt
pymongo
pandas
dnspython
flask